ai/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
anthropic>=0.34
python-dotenv>=1.0
pydantic>=2.0
httpx>=0.27
